+++ go/v1/types/isolation-review.go
@@ -1,4 +1,4 @@
-// Code generated from traust-contracts v0.33.0. DO NOT EDIT.
+// Code generated from traust-contracts v0.36.0. DO NOT EDIT.
 
 package types